requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==3.2.14; python_version < '3.8'
Django==4.0.6 ; python_version >= '3.8'
django-cors-headers==3.13.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3; sys_platform != 'win32'
psycopg2==2.9.3; sys_platform == 'win32'
pytz==2022.1
sqlparse==0.4.2
urllib3==1.26.10
whitenoise==6.2.0
